Save optimization remarks in an external YAML file
This brings the capability from clang to save remarks in an external YAML files.
This allows them to be viewed with tools like the opt-viewer.

Saving the remarks is activated with the new option -fsave-optimization-record.
The command-line interface closely follows the one in clang.

Similarly to -emit-tbd, I've only added support for single-compile mode for now.
In this case the default filename is determined by
getOutputFilenameFromPathArgOrAsTopLevel, i.e. unless explicitly specified
with -foptimization-record-file, the file is placed in the directory of the main
output file as <modulename>.opt.yaml.

I need to think more where the files should end up in non-wmo mode.  Probably
where the .dia files go?  Suggestions are welcome!